move script to scripts, rm datasets
